Get rid of the panic stack trace for the favicon:
If you use a browser to download content from S3Proxy, it results in a
panic stack trace in the S3Proxy server logs. This is caused by Chrome
(possibly others) doing a subsequent request to get the favicon
(/favicon.ico).

Because of the catch-all nature of the regex, S3Proxy tries to lookup a
bucket with that name.

The simple fix is to 404 on the favicon URL